+++ HISTORY.rst
@@ -9,6 +9,8 @@ History
 * stylechecker's `--loglevel` option accepts upper, lower or mixed case strings.
 * stylechecker utility can read from stdin, so it can be a filter in unix 
   pipelines.
+* Added option `--noaggregation` to stylechecker, to prevent output from being 
+  aggregated as a single JSON data structure.
 
 
 0.7.4 (2015-06-19)